+++ src/components/node-health-icon/node-health-icon.e2e.ts
@@ -25,7 +25,9 @@ describe('NodeHealthIcon component should', () => {
 
     expect(root).toEqualHtml(`
       <symbol-node-health-icon node-url="http://localhost:3000">
-        <img src="/resources/down.png" title="Disconnected" class="symbol-node-health-icon" />
+        <mock:shadow-root>
+          <img src="/resources/down.png" title="Disconnected" class="symbol-node-health-icon" />
+        </mock:shadow-root>
       </symbol-node-health-icon>
     `);
   });