+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [v0.1.0][v0.1.0] - May 2020
 
-- initial repository structure
+- Added component `<symbol-node-health-icon>`
+- Added configuration for StencilJS webcomponents compiler
 
 [v0.1.0]: https://github.com/symbol/symbol-components/releases/tag/v0.1.0